Extract markdown reading into dedicated file
